Documentation on ruby land.
Made extensive changes to how documentation is organized. Mainly changed comments, but some method's APIs were altered too. Added :nodoc:'s in lots of files, that's why there are changes in spec/ and ext/.

I've changed the NMatrix::VERSION to a module, following Rails convention.

Finally, almost everything has a comment in it, but there are lots of writing to be made - blas.rb and lapack.rb aren't finished yet and need a better explanation of the arguments for each method. Introduction of hashed parameters in them would be great.

NMatrix/NVector/shortcuts coverage is pretty good.

Fixed a bug in slice_spec