yale_vector_insert: pos no longer required for beginning of rows -- now just looks it up based on the row index provided. Arg order also changed.
I tested this with code that builds the Yale matrix row-by-row
and it filled about 12x faster, a huge improvement. There may still be a
problem, however, as the test code was designed not to require resizes but
one still happened. A spec needs to be written -- need to confirm the
matrix's integrity.

I also renamed the function from yale_vector_insert_by_pos().

This could be used to easily build a pure Ruby row-by-row-initialization
function for Yale -- especially useful for converting a Ruby Hash into a
Yale NMatrix.