Added NVector#logspace() shortcut
This commit adds the logspace() shortcut to the NVector class. Its
functionality is almost identical to the homonymous function in MATLAB.

Simply put, it receives parameters `a`, `b` and `n` and generates a NVector
with `n` values from `10 ** a` to `10 ** b`, logarithmically spaced. For
example:

    u = NVector.logspace(0, 1, 2)
    u.pretty_print # => [1, 10]

There are some updates to other NVector's specs in order to simplify
them.