NMatrix::BLAS.cblas_scal make use of CBLAS
First of all, `xscal` (x in {s,d,c,z}) seems to be a BLAS level 1
function instead of a LAPACK one according to the standard[1]. Thus I
reworded it to reflect reality and added calls to the cblas routines
(which also solves an old FIXME).

I still don't know how everything works in the NMatrix C/C++ world, but
it appears to work correctly.

[1]: http://www.compsci.wm.edu/sciclone/documentation/software/math/ATLAS/blas-report.pdf