Use address rather than value in several conservative registrations.
While doing some debugging for #277, I compiled with NM_GC_CONSERVATIVE defined, and found that the code no longer compiled!  This was due to a number of places where I was unregistering a value instead of a pointer.  I've fixed this by passing addresses instead, which allows nmatrix to compile with the conservative gc flag on.

(One additional address vs. value usage I missed due to not cleaning up properly.)