Fix nmatrix.h to work with C-only code.
This was reported as #324.

When including nmatrix.h without __cplusplus defined, there are some
inconsistencies with the NM_ prefix namespacing, and some places use the
unprefixed versions even in C code.

This commit makes the namespacing consistent by defining internal _NMATRIX and
_STORAGE macros that refer to the appropriate namespaced/non-namespaced version
in the API function declarations, and then undefs them at the end so that we
don't pollute the global namespace.  External code should still use NM_ prefixed versions for C and non-prefixed versions for C++

Test Plan:
- create a mini C extension test project that includes nmatrix.h, verify that it compiles correctly
- in nmatrix: `bundle exec rake clobber && bundle exec rake compile && bundle exec rake spec`